+++ lib/src/ex_bloc.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:bloc/bloc.dart';
 
 import './ex_event.dart';
@@ -5,8 +7,11 @@ import './ex_store.dart';
 
 abstract class ExBloc<Event extends ExEvent, State> extends Bloc<Event, State>
     implements ExEventHandler {
+  StreamSubscription<Event> _subscription;
   ExBloc() : super() {
     ExStore.instance.add<State>(this);
+    _subscription =
+        ExStore.instance.on<Event>().listen((Event event) => this.add(event));
     if (initialEvents.isNotEmpty) {
       initialEvents.forEach((e) {
         ExStore.instance.dispatch(e);
@@ -19,6 +24,7 @@ abstract class ExBloc<Event extends ExEvent, State> extends Bloc<Event, State>
   @override
   Future<void> close() {
     ExStore.instance.remove<State>(this);
+    _subscription.cancel();
     return super.close();
   }