+++ test/ex_bloc_test.dart
@@ -5,13 +5,30 @@ import 'package:test/test.dart';
 
 class MockExStore extends Mock implements ExStore {}
 class MockExBloc extends Mock implements ExBloc {}
+class MyExEvent extends Mock implements ExEvent {}
+
+class ExBlocImpl extends ExBloc<MyExEvent, dynamic> {
+  @override
+  get initialState => null;
+} 
 
 void main() {
-  test('adds one to input values', () {
-    // TODO
-    final store = MockExStore();
-    ExStore.initWith(store);
-    // ExStore.instance.dispatch(event)
-
+  test('wait for event', () async {
+    final eventIn = MyExEvent();
+    final eventOutFuture = ExStore.instance.wait<MyExEvent>();
+    expect(eventOutFuture, completion(equals(eventIn)));
+    ExStore.instance.dispatch(eventIn);
   });
+
+  test('test ex bloc interaction', () async {
+    final eventIn = MyExEvent();
+    final bloc = ExBlocImpl();
+    final eventOutFuture = ExStore.instance.wait<MyExEvent>();
+    expect(eventOutFuture, completion(equals(eventIn)));
+    ExStore.instance.dispatch(eventIn);
+    await untilCalled(eventIn.call(any, any));
+    verify(eventIn.call(ExStore.instance, bloc)).called(1);
+    verifyNoMoreInteractions(eventIn);
+  });
+  
 }